+++ lib/wikimelon/property.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Wikimelon
+  class Property < Resource
+    def datatype
+      @entity&.dig('datatype')
+    end
+  end
+end