+++ lib/wikimelon/request.rb
@@ -1,5 +1,7 @@
 require_relative "faraday" # !! Potential ruby 3.0 difference in module loading? relative differs from Serrano
 require "faraday/follow_redirects"
+require "faraday/retry"
+require "json"
 require_relative "utils"
 
 module Wikimelon
@@ -15,6 +17,7 @@ module Wikimelon
       @url = args[:url]
       @verbose = args[:verbose]
       @query = args[:query]
+      @params = args[:params]
       @limit = args[:limit]
       @offset = args[:offset]
       @options = args[:options] # TODO: not added at wikimelon.rb
@@ -22,13 +25,21 @@ module Wikimelon
 
     def perform
 
-      args = {'query': @query, 'format': 'json'}
-      opts = args.delete_if { |_k, v| v.nil? }
+      opts = @params || {'query': @query, 'format': 'json'}.delete_if { |_k, v| v.nil? }
 
       Faraday::Utils.default_space_encoding = "+"
 
+      retry_max = Wikimelon.retry_max.to_i
+
       conn = Faraday.new(url: @url) do |f|
               f.response :logger if verbose
+              if retry_max > 0
+                f.request :retry,
+                          max: retry_max,
+                          interval: Wikimelon.retry_interval.to_f,
+                          backoff_factor: 2,
+                          retry_statuses: [429, 503]
+              end
               f.use Faraday::WikimelonErrors::Middleware
               f.adapter Faraday.default_adapter
              end
@@ -37,9 +48,10 @@ module Wikimelon
       conn.headers[:user_agent] = make_user_agent
       conn.headers["X-USER-AGENT"] = make_user_agent
 
+      Wikimelon::Throttle.wait!
       res = conn.get(endpoint, opts)
 
-      MultiJson.load(res.body)      
+      JSON.parse(res.body)
     end
   end
 end